Move the config feature unit test logic to configuration.spec
* Removed the config.spec.ts from the e2e suite and moved the logic to configuration.spec.ts unit test
* Restored env vars safely by deleting only the added keys
